rename : 모델 클래스 폴더 위치 재분류
