+++ src/main/java/racingcar/model/entity/Car.java
@@ -1,4 +1,4 @@
-package racingcar.model.domain;
+package racingcar.model.entity;
 
 import racingcar.strategy.Strategy;
