+++ src/main/java/racingcar/model/entity/RacingChance.java
@@ -1,4 +1,4 @@
-package racingcar.model.io;
+package racingcar.model.entity;
 
 import racingcar.utils.ErrorMessage;
 
@@ -48,7 +48,7 @@ public class RacingChance {
         }
     }
 
-    public int get() {
+    public int getValue() {
         return chance;
     }
 }
