+++ src/main/java/racingcar/view/ConsoleOutputView.java
@@ -1,8 +1,6 @@
 package racingcar.view;
 
-import racingcar.model.dto.CarRacingProgress;
-import racingcar.model.io.Winner;
-import racingcar.utils.IOMessage;
+import racingcar.model.dto.RacingResult;
 
 /**
  * 콘솔을 통해 출력을 담당하는 클래스
@@ -10,23 +8,7 @@ import racingcar.utils.IOMessage;
 public class ConsoleOutputView implements OutputView {
 
     @Override
-    public void printProgress(CarRacingProgress carDto) {
-        System.out.println(carDto);
+    public void printRacingResult(RacingResult result) {
+        System.out.println(result);
     }
-
-    @Override
-    public void printStartRace() {
-        System.out.println(IOMessage.RACING_START);
-    }
-
-    @Override
-    public void printLineBreaking() {
-        System.out.println();
-    }
-
-    @Override
-    public void printRaceResult(Winner winner) {
-        System.out.println(IOMessage.WINNER_ANNOUNCEMENT.toString() + winner);
-    }
-
 }
