+++ src/main/java/racingcar/view/OutputView.java
@@ -1,18 +1,11 @@
 package racingcar.view;
 
-import racingcar.model.dto.CarRacingProgress;
-import racingcar.model.io.Winner;
+import racingcar.model.dto.RacingResult;
 
 /**
  * 출력을 담당하는 인터페이스
  */
 public interface OutputView {
 
-    void printProgress(CarRacingProgress car);
-
-    void printRaceResult(Winner winner);
-
-    void printStartRace();
-
-    void printLineBreaking();
+    void printRacingResult(RacingResult result);
 }
